WINGs shared library generation
This patch modifies the linking of the WINGs libraries to restore the creation
of a shared library. wmaker used to do this, but it was dropped around
wmaker-0.90/0.91. The shared so library is needed when compiling and running
the wdm display manager.

Submitted by: Gilbert Ashley
Author: Alexey Voinov <voins@altlinux.ru>
Origin: ALT/Sisyphus Linux